refactor ArrayDistinct and ArrayDupes into one class ArrayUnique
